wrote and incorporated new catfinder() command into predictr
catfinder, short for category finder, looks to see if there are
categories in the test set that do not appear in training. It returns
those categories and the full list of categories in test and training
sets. Note that it is based on the model.frame, and therefore may find
more new categories as the usable data in the models decreases.